requirements.txt
Django>=4.2
djangorestframework>=3.15
requests>=2.31
pandas>=2.2
yfinance>=0.2.44
ta>=0.11.0
python-dateutil>=2.9
